Separate SystemBody from StarSystem; Decouple a lot of includes
