+++ src/Game.h
@@ -7,14 +7,12 @@
 #include "JsonFwd.h"
 #include "galaxy/Galaxy.h"
 #include "gameconsts.h"
-#include "libs.h"
 #include <string>
 
 class SystemPath;
 class GameLog;
 class HyperspaceCloud;
 class Player;
-class ShipController;
 class Space;
 
 namespace Graphics {
