+++ src/GeoSphere.cpp
@@ -2,6 +2,7 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "GeoSphere.h"
+
 #include "GameConfig.h"
 #include "GeoPatch.h"
 #include "GeoPatchContext.h"
@@ -15,7 +16,6 @@
 #include "graphics/Texture.h"
 #include "graphics/TextureBuilder.h"
 #include "graphics/VertexArray.h"
-#include "libs.h"
 #include "perlin.h"
 #include "vcacheopt/vcacheopt.h"
 #include <algorithm>
