+++ src/Player.h
@@ -4,11 +4,8 @@
 #ifndef _PLAYER_H
 #define _PLAYER_H
 
-#include "HyperspaceCloud.h"
 #include "Ship.h"
 #include "ShipCockpit.h"
-#include "galaxy/StarSystem.h"
-#include "libs.h"
 #include "ship/PlayerShipController.h"
 #include <list>
