+++ src/Sfx.cpp
@@ -2,22 +2,22 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "Sfx.h"
+
 #include "FileSystem.h"
 #include "Frame.h"
 #include "GameSaveError.h"
 #include "IniConfig.h"
+#include "Json.h"
 #include "JsonUtils.h"
 #include "Pi.h"
 #include "Ship.h"
 #include "Space.h"
 #include "StringF.h"
-#include "galaxy/StarSystem.h"
 #include "graphics/Drawables.h"
 #include "graphics/Graphics.h"
 #include "graphics/Material.h"
 #include "graphics/Renderer.h"
 #include "graphics/TextureBuilder.h"
-#include "libs.h"
 
 using namespace Graphics;
